Sessions invalidated when requesting certain pages
On a normal request, Chrome accepts encodings gzip, deflate, sdch.  When requesting a media file, it drops the `sdch`.

This triggers Rack 1.5.2's session hijacking alarms and the session is invalidated and users are redirected to the login screen.

Rack 1.5.3 doesn't check the HTTP_ACCEPT_ENCODING so upgrading Rack would resolve this problem but the other two headers could easily cause problems in the future.  I'll move the traffic to SSL to solve the underlying security concerns and for now just stop checking for session hijacking.